+++ example/rsa_cipher.dart
@@ -1,4 +1,3 @@
-import 'package:pointycastle/export.dart';
 import 'package:rsa_cipher/rsa_cipher.dart';
 
 void main() async {
@@ -6,24 +5,12 @@ void main() async {
   final keyPair = RsaCipher().generateKeyPair();
 
   // encode key to pem
-  final publicKeyPem = RsaCipher().keyToPem<RSAPublicKey>(keyPair.publicKey);
-  final privateKeyPem = RsaCipher().keyToPem<RSAPrivateKey>(keyPair.privateKey);
+  final publicKeyPem = keyPair.publicKey.toPem();
+  final privateKeyPem = keyPair.privateKey.toPem();
 
   // decode pem to key
-  final publicKey = RsaCipher().keyFromPem<RSAPublicKey>(publicKeyPem);
-  final privateKey = RsaCipher().keyFromPem<RSAPrivateKey>(privateKeyPem);
-
-  // save pem to file
-  RsaCipher().storeKeyToFile(filePath: '.../public_key.pem', key: publicKey);
-  RsaCipher().storeKeyToFile(filePath: '.../private_key.pem', key: privateKey);
-
-  // get key from file
-  final publicKeyFromFile = RsaCipher().retrieveKeyFromFile<RSAPublicKey>(
-    '.../public_key.pem',
-  );
-  final privateKeyFromFile = RsaCipher().retrieveKeyFromFile<RSAPrivateKey>(
-    '.../private_key.pem',
-  );
+  final publicKey = PublicKey.fromPem(publicKeyPem);
+  final privateKey = PrivateKey.fromPem(privateKeyPem);
 
   // encrypt text
   final cipherText = RsaCipher().encrypt(