+++ lib/rsa_cipher.dart
@@ -1,3 +1,2 @@
-library;
-
 export 'src/rsa_cipher.dart';
+export 'src/models/models.dart';