+++ lib/src/models/key_pair.dart
@@ -1,0 +1,8 @@
+import 'private_key.dart';
+import 'public_key.dart';
+
+class KeyPair {
+  KeyPair({required this.privateKey, required this.publicKey});
+  final PrivateKey privateKey;
+  final PublicKey publicKey;
+}